### 2020-08-21 #2591 @whedon
[PRE REVIEW]: astrosource: automating optical astronomy measurement, calibration and analysis for variable stellar sources from provided photometry
**Submitting author:** @mfitzasp (<a href="http://orcid.org/0000-0001-6554-1826">Michael Fitzgerald</a>)
**Repository:** <a href="https://github.com/zemogle/astrosource/" target ="_blank">https://github.com/zemogle/astrosource/</a>
**Version:** v1.4.0
