+++ astrosource/comparison.py
@@ -345,7 +345,7 @@ def catalogue_call(avgCoord, opt, cat_name, targets, closerejectd):
     data.emag = array(resp[opt['error']].data)
     return data
 
-def calibrate_photometry(targets, filterCode, compFile, closerejectd, variabilityMultiplier, nopanstarrs=False, nosdss=False):
+def calibrate_photometry(targets, filterCode, starvar, closerejectd, variabilityMultiplier, nopanstarrs=False, nosdss=False):
     FILTERS = {
                 'B' : {'APASS' : {'filter' : 'Bmag', 'error' : 'e_Bmag'}},
                 'V' : {'APASS' : {'filter' : 'Vmag', 'error' : 'e_Vmag'}},
@@ -370,14 +370,14 @@ def calibrate_photometry(targets, filterCode, compFile, closerejectd, variabilit
 
     logger.debug("Filter Set: " + filterCode)
 
-    logger.debug(compFile.shape[0])
+    logger.debug(starvar.shape[0])
 
-    compCoords = SkyCoord(ra=compFile[:,0]*degree, dec=compFile[:,1]*degree)
+    compCoords = SkyCoord(ra=starvar[:,0]*degree, dec=starvar[:,1]*degree)
 
     # Get Average RA and Dec from file
-    logger.debug(average(compFile[:,0]))
-    logger.debug(average(compFile[:,1]))
-    avgCoord = SkyCoord(ra=(average(compFile[:,0]))*degree, dec=(average(compFile[:,1]))*degree)
+    logger.debug(average(starvar[:,0]))
+    logger.debug(average(starvar[:,1]))
+    avgCoord = SkyCoord(ra=(average(starvar[:,0]))*degree, dec=(average(starvar[:,1]))*degree)
 
     try:
         catalogues = FILTERS[filterCode]
@@ -418,7 +418,7 @@ def calibrate_photometry(targets, filterCode, compFile, closerejectd, variabilit
     # So the stars that will be used to calibrate the frames to get the OTHER stars.
     calibStands=[]
 
-    for compstar in compFile:
+    for compstar in starvar:
         # Match each potential comparison star to star in online catalogue
         compCoord=SkyCoord(ra=compstar[0]*degree, dec=compstar[1]*degree)
         idxcomp,d2dcomp,d3dcomp=compCoord.match_to_catalog_sky(catCoords)
@@ -447,10 +447,11 @@ def calibrate_photometry(targets, filterCode, compFile, closerejectd, variabilit
     return calibStands, goodcalib, cat_used
 
 
-def find_comparisons_calibrated(targets, paths, filterCode, photlist, comparisons, compFile, nopanstarrs=False, nosdss=False, closerejectd=5.0, max_magerr=0.05, stdMultiplier=2, variabilityMultiplier=2):
+def find_comparisons_calibrated(targets, paths, filterCode, photlist, comparisons, starvar, nopanstarrs=False, nosdss=False, closerejectd=5.0, max_magerr=0.05, stdMultiplier=2, variabilityMultiplier=2):
     sys.stdout.write("⭐️ Find comparison stars in catalogues for calibrated photometry\n")
-
-    calibStands, goodcalib, cat_used = calibrate_photometry(targets=targets, filterCode=filterCode, compFile=compFile, closerejectd=closerejectd, nopanstarrs=nopanstarrs, nosdss=nosdss, variabilityMultiplier=variabilityMultiplier)
+    compsused = photlist[0][comparisons][:,[0,1]]
+    logger.critical(compsused)
[…]
