### 2020-08-22 aa18186
fixing up strreject/imgreject routine for small numbers of stars
