### 2020-08-22 e9b89bf
tiny fix to displaying what number variable was being processed
