### 2020-08-22 c943b79
small glitch when only one comparison star in comparison.py
