+++ joss/paper.bib
@@ -22,6 +22,7 @@ archivePrefix = {arXiv},
   journal={Publications of the Astronomical Society of Australia},
   volume={31},
   year={2014},
+  doi = {https://doi.org/10.1017/pasa.2014.30},
   publisher={Cambridge University Press}
 }
 
@@ -33,6 +34,7 @@ archivePrefix = {arXiv},
   number={3},
   pages={239--245},
   year={2016},
+  doi = {https://doi.org/10.1002/asna.201512254},
   publisher={Wiley Online Library}
 }
 
@@ -43,7 +45,8 @@ archivePrefix = {arXiv},
   volume={153},
   number={2},
   pages={77},
-  year={2017},
+  year={2017},  
+  doi = {https://doi.org/10.3847/1538-3881/153/2/77},
   publisher={IOP Publishing}
 }
 
@@ -55,6 +58,7 @@ archivePrefix = {arXiv},
   number={1011},
   pages={054401},
   year={2020},
+  doi = {https://doi.org/10.1088/1538-3873/ab7ee7},
   publisher={IOP Publishing}
 }
 
@@ -80,6 +84,7 @@ archivePrefix = {arXiv},
   journal={PASA},
   volume={35},
   pages={e010},
+  doi = {https://doi.org/10.1017/pasa.2018.5},
   year={2018}
 }
 
@@ -91,6 +96,7 @@ archivePrefix = {arXiv},
   number={1},
   pages={12},
   year={2015},
+  doi = {https://doi.org/10.1088/0067-0049/219/1/12},
   publisher={IOP Publishing}
 }
 
@@ -109,6 +115,7 @@ archivePrefix = {arXiv},
   journal={Astronomy Theory, Observations \& Methods},
   volume={1},
   number={1},
+  doi = {https://doi.org/10.32374/atom.2020.1.2},
   year={2020}
 }
 
@@ -118,6 +125,7 @@ archivePrefix = {arXiv},
   journal={Astronomy Theory, Observations \& Methods},
   volume={1},
   number={1},
+  doi = {https://doi.org/10.32374/atom.2020.1.4},
   year={2020}
 }
 
@@ -220,6 +228,7 @@ Bdsk-Url-1 = {https://doi.org/10.3847/1538-3881/aabc4f}}
   year={2011},
   volume={13},
   number={2},
+  doi = {https://doi.org/10.1109/mcse.2011.37},
   pages={22-30},}
 
 @ARTICLE{matplotlib,
@@ -229,6 +238,7 @@ Bdsk-Url-1 = {https://doi.org/10.3847/1538-3881/aabc4f}}
   year={2007},
   volume={9},
   number={3},
+  doi = {https://doi.org/10.1109/mcse.2007.55},
   pages={90-95},}
 
 @ARTICLE{astroquery,
@@ -277,6 +287,7 @@ Bdsk-Url-1 = {https://doi.org/10.3847/1538-3881/aabc4f}}
   journal={Astronomy Theory, Observations \& Methods},
   volume={1},
   number={1},
+  doi = {https://doi.org/10.32374/atom.2020.1.1},
   year={2020}
 }
 
@@ -288,5 +299,6 @@ Bdsk-Url-1 = {https://doi.org/10.3847/1538-3881/aabc4f}}
   number={1},
   pages={369--377},
   year={2002},
+  doi = {https://doi.org/10.1051/0004-6361:20020802},
   publisher={EDP Sciences}
 }
