### 2020-08-23 e819b03
updating readme.md and docs
