+++ README.md
@@ -60,8 +60,20 @@ There are a few input options when running the scripts. You can either run the w
 
 `--full` [boolean flag] Run the whole code. This will run the following steps in this order `stars` > `comparison` > `calc` > `phot` > `plot`
 
+
+
 ### Extra options
 `--verbose` [boolean flag] Show all system messages for AstroSource
+
+`--period` [boolean flag] Search for periodicity in the data, currently with PDM and String methods. This will autoselect a reasonable search range if not provided a range.
+
+`--periodlower` [float] Shortest period to trial in days. Default is 0.05
+
+`--periodupper` [float] Longest period to trial in days. Default is one-third the observational baseline of your dataset.
+
+`--periodtests` [integer] Number of different trial periods to run, Default 10000
+
+`--skipvarsearch` [boolean flag] If this is set, this skips the variability calculations for each identified star. Pragmatically this skips creating the starVariability outputs. In a crowded field this can take an excessive amount of time. 
 
 `--detrend` [boolean flag] Detrend exoplanet data
 
@@ -69,9 +81,25 @@ There are a few input options when running the scripts. You can either run the w
 
 `--calib` [boolean flag] Perform calibrated
 
-`--imgreject` [float] Image fraction rejection allowance. Defaults to `0.0`. Increasing this will allow AstroSource to reject some of your data files if there are not enough comparison stars.
+`--lowcounts` [int] Countrate above which to accept a comparison star as a candidate. Defaults to 1000. (Note: This may seem like a low number but realistically stars at this low rate will be highly variable and rejected anyway)
+
+`--hicounts` [int] Countrate above which to reject a comparison star as a candidate. Defaults to 1000000. (Note: This will vary from camera to camera, but it should be representative of a typical value that would have a peak pixel value significantly below the full range of the ccd, preferably lower rather than higher. )
+
+`--thresholdcounts` [int] the number of counts at which to stop adding identified comparison stars to the ensemble. Default 1000000. 
+
+`--closerejectd` [float] astrosource will reject potential comparison calibration stars if there are nearby stars to it. Closerejectd is the limiting distance in arcseconds. For crowded fields, this value may need to be lowered. Default 5.0. While the primary function here is to identify stars that are not in crowded situations, it also has the side-effect of removing false detections in sky surveys due to image artifacts or diffraction spikes which tend to cluster together.
+
+`--nopanstarrs` [boolean flag] Do not use the PanSTARRS catalogue for calibration. Some regions of the sky in PanSTARRS have poorer quality photometry than SDSS.
+
+`--sdss` [boolean flag] Do not use the SDSS catalogue for calibration. Some regions of the sky in SDSS have poorer quality photometry than PanSTARRS.
+
+`--bjd` [boolean flag] Convert the MJD time into BJD time for LCO images.
 
 `--clean` [boolean flag] Remove all files except the original data files, and photometry files
+
+`--imgreject` [float] Image fraction rejection allowance based on image size starting value. Defaults to `0.0`. Astrosource automatically adjusts this value, so it is only in very rare cases this might need to be set.
+
+`--starreject` [float] Image fraction rejection allowance based on number of rejected stars starting value. Defaults to `0.3`. Astrosource automatically adjusts this value, so it is only in very rare cases this might need to be set.
 
 
 ### Example Usage
@@ -81,6 +109,41 @@ astrosource --ra 154.9083708 --dec -9.8062778 --indir /path/to/your/data --full
 ```
 
 All the files generated will be stored in the directory you specify in `--indir`
+
+### Output Files
+
+A variety of output files are generated. Some are obvious, some are not so obvious.
+
+`usedImages.txt` : the list of images astrosource chose to use out of the original image set
+
[…]
