### 2020-08-23 d171d10
glitch - calibcats was meant to output csv, not npy
