+++ astrosource/test/test_find_stars.py
@@ -1,0 +1,143 @@
+import numpy as np
+from pathlib import Path
+import os
+import pytest
+from unittest.mock import patch, Mock
+
+from astrosource.identify import find_stars, read_data_files
+from astrosource.comparison import find_comparisons, find_comparisons_calibrated, \
+    calibrate_photometry
+
+TEST_PATH_PARENT = Path(os.path.dirname(__file__)) / 'test_files'
+
+TEST_PATHS = {'parent': TEST_PATH_PARENT / 'speedup'}
+
+INPUT_FILES = ['RXEri_B_2458372d8045823071_2018d09d11T07d09d16d123_1a3954565_17d286_kb81_michaelfitzgerald.npy',
+ 'RXEri_B_2458547d2593600489_2019d03d04T18d20d17d106_1a1375401_13d283_kb84_timdjones7.npy',
+ 'RXEri_B_2458549d2488945760_2019d03d06T18d05d17d487_1a124071_13d284_kb84_timdjones7.npy',
+ 'rxeri_B_2458562d2884747311_2019d03d19T19d02d33d828_1a5178674_13d285_kb84_timdjones7.npy',
+ 'RXEri_B_2458372d1298749419_2018d09d10T14d57d43d008_1a2899937_16d942_kb27_michaelfitzgerald.npy',
+ 'RXEri_B_2458371d6046093008_2018d09d10T02d21d20d615_1a1529288_17d283_kb96_michaelfitzgerald.npy',
+ 'RXEri_B_2458371d8837984502_2018d09d10T09d03d22d513_1a0790084_17d284_kb26_michaelfitzgerald.npy',
+ 'RXEri_B_2458546d2483008588_2019d03d03T18d04d18d732_1a1029248_13d409_kb84_timdjones7.npy',
+ 'rxeri_B_2458567d2286917502_2019d03d24T17d36d28d604_1a2230736_13d284_kb84_timdjones7.npy',
+ 'RXEri_B_2458371d6050731102_2018d09d10T02d22d00d686_1a1514762_17d282_kb96_michaelfitzgerald.npy',
+ 'RXEri_B_2458371d7887208089_2018d09d10T06d46d27d791_1a5547859_17d285_kb26_michaelfitzgerald.npy',
+ 'RXEri_B_2458372d2446666160_2018d09d10T17d43d00d539_1a153668_17d238_kb24_michaelfitzgerald.npy',
+ 'RXEri_B_2458371d6041340921_2018d09d10T02d20d40d435_1a1543828_17d283_kb96_michaelfitzgerald.npy',
+ 'RXEri_B_2458550d2448007651_2019d03d07T17d59d26d062_1a1206428_13d284_kb84_timdjones7.npy',
+ 'RXEri_B_2458553d3472028011_2019d03d10T20d26d58d726_1a5740949_13d338_kb25_timdjones7.npy',
+ 'rxeri_B_2458563d4888034821_2019d03d20T23d51d02d943_1a1935013_13d284_kb95_timdjones7.npy',
+ 'RXEri_B_2458556d7220212580_2019d03d14T05d26d47d877_1a3792819_12d941_kb82_timdjones7.npy',
+ 'rxeri_B_2458564d2591135348_2019d03d21T18d20d18d252_1a3380838_13d286_kb84_timdjones7.npy',
+ 'rxeri_B_2458572d4780848781_2019d03d29T23d35d31d896_1a2556188_13d27_kb95_timdjones7.npy',
+ 'RXEri_B_2458371d7881178972_2018d09d10T06d45d36d244_1a5607974_17d282_kb26_michaelfitzgerald.npy',
[…]
