+++ joss/paper.bib
@@ -22,7 +22,7 @@ archivePrefix = {arXiv},
   journal={Publications of the Astronomical Society of Australia},
   volume={31},
   year={2014},
-  doi = {https://doi.org/10.1017/pasa.2014.30},
+  doi = {10.1017/pasa.2014.30},
   publisher={Cambridge University Press}
 }
 
@@ -34,7 +34,7 @@ archivePrefix = {arXiv},
   number={3},
   pages={239--245},
   year={2016},
-  doi = {https://doi.org/10.1002/asna.201512254},
+  doi = {10.1002/asna.201512254},
   publisher={Wiley Online Library}
 }
 
@@ -46,7 +46,7 @@ archivePrefix = {arXiv},
   number={2},
   pages={77},
   year={2017},  
-  doi = {https://doi.org/10.3847/1538-3881/153/2/77},
+  doi = {10.3847/1538-3881/153/2/77},
   publisher={IOP Publishing}
 }
 
@@ -58,7 +58,7 @@ archivePrefix = {arXiv},
   number={1011},
   pages={054401},
   year={2020},
-  doi = {https://doi.org/10.1088/1538-3873/ab7ee7},
+  doi = {10.1088/1538-3873/ab7ee7},
   publisher={IOP Publishing}
 }
 
@@ -84,7 +84,7 @@ archivePrefix = {arXiv},
   journal={PASA},
   volume={35},
   pages={e010},
-  doi = {https://doi.org/10.1017/pasa.2018.5},
+  doi = {10.1017/pasa.2018.5},
   year={2018}
 }
 
@@ -96,7 +96,7 @@ archivePrefix = {arXiv},
   number={1},
   pages={12},
   year={2015},
-  doi = {https://doi.org/10.1088/0067-0049/219/1/12},
+  doi = {10.1088/0067-0049/219/1/12},
   publisher={IOP Publishing}
 }
 
@@ -115,7 +115,7 @@ archivePrefix = {arXiv},
   journal={Astronomy Theory, Observations \& Methods},
   volume={1},
   number={1},
-  doi = {https://doi.org/10.32374/atom.2020.1.2},
+  doi = {10.32374/atom.2020.1.2},
   year={2020}
 }
 
@@ -125,7 +125,7 @@ archivePrefix = {arXiv},
   journal={Astronomy Theory, Observations \& Methods},
   volume={1},
   number={1},
-  doi = {https://doi.org/10.32374/atom.2020.1.4},
+  doi = {10.32374/atom.2020.1.4},
   year={2020}
 }
 
@@ -203,7 +203,7 @@ Primaryclass = {astro-ph.IM},
 Title = {{Astropy: A community Python package for astronomy}},
 Volume = 558,
 Year = 2013,
-Bdsk-Url-1 = {https://dx.doi.org/10.1051/0004-6361/201322068}}
+Bdsk-Url-1 = {10.1051/0004-6361/201322068}}
 
 @article{astropy:2018,
 Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
@@ -219,7 +219,7 @@ Primaryclass = {astro-ph.IM},
 Title = {{The Astropy Project: Building an Open-science Project and Status of the v2.0 Core Package}},
 Volume = {156},
 Year = 2018,
-Bdsk-Url-1 = {https://doi.org/10.3847/1538-3881/aabc4f}}
+Bdsk-Url-1 = {10.3847/1538-3881/aabc4f}}
 
 @ARTICLE{numpy,
   author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},
[…]
