### 2020-08-23 589a282
Made sanityplot csv and pngs
