### 2020-08-24 69662ed
Made sanityplot least squares fits, plot and output coefficients
