### 2020-08-24 ff8df82
Corrected photometry
