+++ astrosource/comparison.py
@@ -64,25 +64,29 @@ def find_comparisons(targets, parentPath=None, fileList=None, photlist=[], stdMu
     logger.debug("Total total {}".format(np.sum(np.array(fileCount))))
     starRejecter = [True for i in range(comparisons.shape[1])]
     oldrejects = [False for i in range(comparisons.shape[1])]
+    stdCompStar, sortStars = calculate_comparison_variation(comparisons, fileCount)
+
     while True:
 
         # Calculate the variation in each candidate comparison star in brightness
         # compared to this gigantic comparison star.
-        numfiles = comparisons[0].size
-        stdCompStar, sortStars = calculate_comparison_variation(comparisons, fileCount)
-
-        variabilityMax=(min(stdCompStar)*variabilityMultiplier)
+        stablestarstd = stdCompStar[starRejecter]
+        variabilityMax=(min(stablestarstd)*variabilityMultiplier)
 
         # Calculate and present the sample statistics
-        stdCompMed = median(stdCompStar)
-        stdCompStd = std(stdCompStar)
+        stdCompMed = median(stablestarstd)
+        stdCompStd = std(stablestarstd)
 
-        logger.debug(f"Median of comparisons = {stdCompMed}")
-        logger.debug(f"STD of comparisons = {stdCompStd}")
+        # Calculate and present statistics of sample of candidate comparison stars.
+        logger.info("Median variability {:.6f}".format(stdCompMed))
+        logger.info("Std variability {:.6f}".format(stdCompStd))
+        logger.info("Min variability {:.6f}".format(min(stablestarstd)))
+        logger.info("Max variability {:.6f}".format(max(stablestarstd)))
+        logger.info("Number of Stable Comparison Candidates {}".format(sum(starRejecter)))
 
         # Delete comparisons that have too high a variability
-        if min(stdCompStar) > 0.002:
-            for j, stdComp in enumerate(stdCompStar):
+        if min(stablestarstd) > 0.002:
+            for j, stdComp in enumerate(stablestarstd):
                 #logger.debug(stdCompStar[j])
                 if stdComp > (stdCompMed + (stdMultiplier*stdCompStd)) or isnan(stdComp):
                     logger.debug(f"Star {j} Rejected, Variability too high")
@@ -94,15 +98,6 @@ def find_comparisons(targets, parentPath=None, fileList=None, photlist=[], stdMu
         else:
             logger.info("Minimum variability is too low for comparison star rejection by variability.")
 
-
-        comparisons = array([c[starRejecter] for c in comparisons])
-
-        # Calculate and present statistics of sample of candidate comparison stars.
-        logger.info("Median variability {:.6f}".format(median(stdCompStar)))
-        logger.info("Std variability {:.6f}".format(std(stdCompStar)))
-        logger.info("Min variability {:.6f}".format(min(stdCompStar)))
-        logger.info("Max variability {:.6f}".format(max(stdCompStar)))
-        logger.info("Number of Stable Comparison Candidates {}".format(comparisons.shape[1]))
         # Once we have stopped rejecting stars, this is our final candidate catalogue then we start to select the subset of this final catalogue that we actually use.
         if starRejecter == oldrejects :
             break
@@ -111,6 +106,11 @@ def find_comparisons(targets, parentPath=None, fileList=None, photlist=[], stdMu
             sys.stdout.write('💫')
             sys.stdout.flush()
             oldrejects[:] = starRejecter
+
+        logger.critical(len(starRejecter))
[…]
