### 2020-08-24 3f14fc5
Removing debug messages
