### 2020-08-25 ec96f0a
Fixed crash on no residuals in sanityplot
