### 2020-08-25 #2591 @zemogle
Hmmm. Those are the DOIs referenced in ATOM. It looks like they don't resolve correctly.
