### 2020-08-25 #2591 @arfon
> Hmmm. Those are the DOIs referenced in ATOM. It looks like they don't resolve correctly.

Indeed. I've enquired here: https://twitter.com/arfon/status/1298252048861814784
