### 2020-08-25 #2591 @arfon
👍 thanks for the quick response @rzellem! :wave: @pearsonkyle - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
