### 2020-08-25 #2591 @bsipocz
@arfon - what is the expected timeline? I'm fully occupied this week and next with AstroHackWeek, but would be happy to start with the review afterwards.
