### 2020-08-26 #2591 @arfon
> @arfon - what is the expected timeline? I'm fully occupied this week and next with AstroHackWeek, but would be happy to start with the review afterwards.

Ideally sometime in the next ~4 weeks?
