### 2020-08-26 9790829
Speedup on photometry
