+++ astrosource/comparison.py
@@ -22,7 +22,7 @@ import logging
 logger = logging.getLogger('astrosource')
 
 
-def find_comparisons(targets, parentPath=None, fileList=None, photlist=[], stdMultiplier=2.5, thresholdCounts=1000000, variabilityMultiplier=2.5, removeTargets=True, acceptDistance=1.0):
+def find_comparisons(targets, parentPath=None, fileList=None, photometry=[], stdMultiplier=2.5, thresholdCounts=1000000, variabilityMultiplier=2.5, removeTargets=True, acceptDistance=1.0):
     '''
     Find stable comparison stars for the target photometry
 
@@ -54,17 +54,17 @@ def find_comparisons(targets, parentPath=None, fileList=None, photlist=[], stdMu
     if type(parentPath) == 'str':
         parentPath = Path(parentPath)
 
-    comparisons = remove_stars_targets(photlist, acceptDistance, targets, removeTargets)
+    photometry = remove_stars_targets(photometry, acceptDistance, targets, removeTargets)
 
     # Add up all of the counts of all of the comparison stars
     # To create a gigantic comparison star.
     logger.debug("Please wait... calculating ensemble comparison star for each image")
 
-    fileCount = np.ndarray.sum(comparisons, axis=1)[:,4]
+    fileCount = np.ndarray.sum(photometry, axis=1)[:,4]
     logger.debug("Total total {}".format(np.sum(np.array(fileCount))))
-    starRejecter = [True for i in range(comparisons.shape[1])]
-    oldrejects = [False for i in range(comparisons.shape[1])]
-    stdCompStar, sortStars = calculate_comparison_variation(comparisons, fileCount)
+    starRejecter = [True for i in range(photometry.shape[1])]
+    oldrejects = [False for i in range(photometry.shape[1])]
+    stdCompStar, starvar = calculate_comparison_variation(photometry, fileCount)
 
     while True:
 
@@ -84,7 +84,7 @@ def find_comparisons(targets, parentPath=None, fileList=None, photlist=[], stdMu
         logger.info("Max variability {:.6f}".format(max(stablestarstd)))
         logger.info("Number of Stable Comparison Candidates {}".format(sum(starRejecter)))
 
-        # Delete comparisons that have too high a variability
+        # Delete photometry that have too high a variability
         if min(stablestarstd) > 0.002:
             for j, stdComp in enumerate(stablestarstd):
                 #logger.debug(stdCompStar[j])
@@ -107,21 +107,20 @@ def find_comparisons(targets, parentPath=None, fileList=None, photlist=[], stdMu
             sys.stdout.flush()
             oldrejects[:] = starRejecter
 
-    comparisons = array([c[starRejecter] for c in comparisons])
-    sortStars = sortStars[starRejecter]
+    photometry = array([c[starRejecter] for c in photometry])
+    starvar = starvar[starRejecter]
 
     sys.stdout.write('\n')
     logger.info('Statistical stability reached.')
     # Sort through and find the largest file and use that as the reference file
-    final_comparisons = final_candidate_catalogue(parentPath, comparisons, sortStars, thresholdCounts, variabilityMax)
-    # return comparisons[:,final_comparisons,:]
-    return comparisons, final_comparisons, sortStars
+    final_comparisons = final_candidate_catalogue(parentPath, photometry, starvar, thresholdCounts, variabilityMax)
+    return photometry, final_comparisons, starvar
 
[…]
