+++ astrosource/analyse.py
@@ -43,7 +43,7 @@ def get_total_counts(photFileArray, compFile, loopLength):
         allCountsArray.append([allCounts, allCountsErr])
     return allCountsArray
 
-def find_variable_stars(targets, comparisons, acceptDistance=1.0, errorReject=0.05, parentPath=None):
+def find_variable_stars(targets, photometry, acceptDistance=1.0, errorReject=0.05, parentPath=None):
     '''
     Find stable comparison stars for the target photometry and remove variables
 
@@ -65,18 +65,38 @@ def find_variable_stars(targets, comparisons, acceptDistance=1.0, errorReject=0.
     minimumVariableCounts = 10000  # Do not try to detect variables dimmer than this.
     minimumNoOfObs = 10 # Minimum number of observations to count as a potential variable.
 
-    photFileArray = comparisons[0:]
+
+    # Load in list of used files
+    fileList = []
+    with open(parentPath / "usedImages.txt", "r") as f:
+        for line in f:
+            fileList.append(line.strip())
+
+    # LOAD Phot FILES INTO LIST
+    photFileArray = photometry[1:]
+
+    # LOAD IN COMPARISON FILE
+    preFile = genfromtxt(parentPath / 'stdComps.csv', dtype=float, delimiter=',')
+
+    if preFile.shape[0] != 13:
+        preFile=(preFile[preFile[:, 2].argsort()])
+
+    # GET REFERENCE IMAGE
+    # Sort through and find the largest file and use that as the reference file
+    referenceFrame = photometry[0]
+
     compFile = genfromtxt(parentPath / "compsUsed.csv", dtype=float, delimiter=',')
     logger.debug("Stable Comparison Candidates below variability threshold")
     outputPhot = []
 
     # Get total counts for each file
 
-    allCountsArray = get_total_counts(photFileArray, compFile, loopLength=compFile.shape[0])
+    allCountsArray = get_total_counts(photometry, compFile, loopLength=compFile.shape[0])
+
 
     # Define targetlist as every star in referenceImage above a count threshold
     logger.debug("Setting up Variable Search List")
-    targetFile = comparisons[0]
+    targetFile = referenceFrame
     # Although remove stars that are below the variable countrate
     starReject=[]
     for q in range(targetFile.shape[0]):
@@ -146,53 +166,27 @@ def find_variable_stars(targets, comparisons, acceptDistance=1.0, errorReject=0.
 
     return outputVariableHolder
 
-def photometric_calculations(targets, comparisons, paths, acceptDistance=5.0, errorReject=0.5, filesave=True):
+def photometric_calculations(targetphot, photometry, paths, fileList, calibrated=False, acceptDistance=5.0, errorReject=0.5, filesave=True):
     fileCount=[]
     photometrydata = []
     sys.stdout.write('🖥 Starting photometric calculations\n')
 
-    photFileArray = comparisons
+    # allCountsArray = get_total_counts(photometry, compFile, loopLength)
 
-    if (paths['parent'] / 'calibCompsUsed.csv').exists():
-        logger.debug("Calibrated")
-        compFile=genfromtxt(paths['parent'] / 'calibCompsUsed.csv', dtype=float, delimiter=',')
-        calibFlag=1
-    else:
-        logger.debug("Differential")
-        compFile=genfromtxt(paths['parent'] / 'compsUsed.csv', dtype=float, delimiter=',')
-        calibFlag=0
-
-    # Get total counts for each file
-    if compFile.shape[0]== 5 and compFile.size ==5:
-        loopLength=1
-    else:
-        loopLength=compFile.shape[0]
-    allCountsArray = get_total_counts(photFileArray, compFile, loopLength)
-
[…]

[…]
