### 2020-08-26 0e91202
Travis badge and version bump
