### 2020-08-26 #2591 @arfon
@bsipocz - I'm going to take your 👍 to [my comment](https://github.com/openjournals/joss-reviews/issues/2591#issuecomment-680683904) above as consent to add you as a reviewer. Please let me know if this isn't correct!
