### 2020-09-01 8066486
Final array conversion
