+++ astrosource/analyse.py
@@ -172,14 +172,12 @@ def photometric_calculations(targetphot, photometry, paths, fileList, calibrated
     sys.stdout.write('🖥 Starting photometric calculations\n')
 
     # allCountsArray = get_total_counts(photometry, compFile, loopLength)
+    n, x = np.array(fileList).shape
+    photsum = np.sum(photometry, axis=1)
 
-    photsum = np.sum(photometry, axis=0)
-    logger.critical(countsum.shape)
-    sys.exit()
     allcountscount=0
-
     # For each target calculate all the things
-    for q, target in enumerate(targets):
+    for q, target in enumerate([targetphot[0]]):
         starErrorRejCount=0
         starDistanceRejCount=0
         logger.debug("****************************")
@@ -194,127 +192,70 @@ def photometric_calculations(targetphot, photometry, paths, fileList, calibrated
         compArray=[]
         compList=[]
         allcountscount=0
-        for imgs, photFile in enumerate(photometry):
-            sys.stdout.write('.')
-            compList=[]
-            fileRaDec = SkyCoord(ra=photFile[:,0]*degree, dec=photFile[:,1]*degree)
-            idx, d2d, _ = varCoord.match_to_catalog_sky(fileRaDec)
-            starRejected=0
-            if (less(d2d.arcsecond, acceptDistance)):
-                magErrVar = 1.0857 * (photFile[idx][5]/photFile[idx][4])
-                if magErrVar < errorReject:
 
-                    magErrEns = 1.0857 * (allCountsArray[allcountscount][1]/allCountsArray[allcountscount][0])
-                    magErrTotal = pow( pow(magErrVar,2) + pow(magErrEns,2),0.5)
+        targetmagerror = 1.0857 * targetphot[:,5]/targetphot[:,4]
+        magerrorens = 1.0857 * photsum[:,5]/photsum[:,4] #magerens
+        magerrtotal = pow(pow(targetmagerror, 2) + pow(magerrorens, 2), 0.5)
+        absmag = 2.5 * log10(photsum[:,4]/targetphot[:,4])
+        logger.critical(photsum[:,4])
+        logger.critical(targetphot[:,4])
+        logger.critical(absmag)
+        dates = np.array(fileList)[:,1].reshape(n,1)
+        airmass = np.array(fileList)[:,2].reshape(n,1)
 
-                    #templist is a temporary holder of the resulting file.
-                    tempList=photFile[idx,0:6]
-                    # logger.debug(f"{tempList}")
-                    googFile = Path(fileList[imgs]).name
-                    tempList = append(tempList, float(googFile.split("_")[2].replace("d",".")))
-                    tempList = append(tempList, float(googFile.split("_")[4].replace("a",".")))
-                    tempList = append(tempList, allCountsArray[allcountscount][0])
-                    tempList = append(tempList, allCountsArray[allcountscount][1])
+        validfiles = targetmagerror < errorReject
 
-                    #Differential Magnitude
-                    tempList = append(tempList, 2.5 * log10(allCountsArray[allcountscount][0]/photFile[idx][4]))
-                    tempList = append(tempList, magErrTotal)
-                    tempList = append(tempList, photFile[idx][4])
-                    tempList = append(tempList, photFile[idx][5])
-
-                    if (compFile.shape[0]== 5 and compFile.size ==5) or (compFile.shape[0]== 3 and compFile.size ==3):
-                        loopLength=1
-                    else:
-                        loopLength=compFile.shape[0]
-                    for j in range(loopLength):
-                        if compFile.size == 2 or (compFile.shape[0]== 3 and compFile.size ==3) or (compFile.shape[0]== 5 and compFile.size ==5):
[…]

[…]
