### 2020-09-03 4634834
Fixed odd list item removal
