+++ .github/workflows/publish-to-test-pypi.yml
@@ -1,3 +1,7 @@
+name: Publish 🚀 astrosource distributions 📦 to PyPI and TestPyPI
+
+on: push
+
 jobs:
   build-n-publish:
     name: Build and publish astrosource package 📦 to PyPI and TestPyPI
