+++ .github/workflows/publish-to-test-pypi.yml
@@ -1,6 +1,9 @@
 name: Publish 🚀 astrosource distributions 📦 to PyPI and TestPyPI
 
-on: push
+on:
+  push:
+    tags:
+      - "*"
 
 jobs:
   build-n-publish:
@@ -31,3 +34,8 @@ jobs:
       with:
         password: ${{ secrets.test_pypi_password }}
         repository_url: https://test.pypi.org/legacy/
+    - name: Publish distribution 📦 to PyPI
+      if: startsWith(github.ref, 'refs/tags')
+      uses: pypa/gh-action-pypi-publish@master
+      with:
+        password: ${{ secrets.pypi_password }}
