+++ .github/workflows/publish-to-test-pypi.yml
@@ -1,9 +1,6 @@
 name: Publish 🚀 astrosource distributions 📦 to PyPI and TestPyPI
 
-on:
-  push:
-    tags:
-      - "*"
+on: push
 
 jobs:
   build-n-publish:
