### 2020-09-05 #2591 @arfon
:wave: @joshspeagle @catherinezucker - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
