### 2020-09-05 #2641 @whedon
[REVIEW]: astrosource: automating optical astronomy measurement, calibration and analysis for variable stellar sources from provided photometry
**Submitting author:** @mfitzasp (<a href="http://orcid.org/0000-0001-6554-1826">Michael Fitzgerald</a>)
**Repository:** <a href="https://github.com/zemogle/astrosource/" target ="_blank">https://github.com/zemogle/astrosource/</a>
**Version:** v1.5.2
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.14113667.v1" target="_blank">10.6084/m9.figshare.14113667.v1</a>
