### 2020-09-06 a8dfa0c
fixed issue where comparisons are too variable -- iteratively increase the varmin to compensate
