### 2020-09-07 292619c
a much simpler, less dumb way to fix the varimin issue
