### 2020-09-07 969cc57
Bug fix for highly variable stars
