### 2020-09-09 6313693
minor tweaks to varimin
