### 2020-09-10 #2641 @bsipocz
@arfon - in this the place to ask paper related questions (rather than software related)?

E.g. authorship related questions should be discussed here, for example the explicit opt in from authors, or in an issue in the repo? (I would this here is better, but there may be a guideline saying otherwise that I missed). Also, I didn't see any discussion in the guidelines about authors for whom there is no trace of contribution in the software repo. Or should the commit history on be used to determined the lead author has made significant contribution to the software?
