### 2020-09-10 #2641 @bsipocz
> ... active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed...

@mfitzasp - Since you and @zemoge interacted with the JOSS submission I take it as your agreement to be authors. I don't see any interactions in the repo with the other two authors though, have they been actively agreed to be listed?

Also, I wonder whether you considered adding the third person as author, who contributing to the package, especially early on helping with setting up the package skeleton, CI, etc. Unless they fully opted out, I would appreciate if they were at least mentioned in the acknowledgement.
