### 2020-09-11 #2641 @zemogle
That's a really good call. @joesingo would you like to be an author? You helped substantially with  the `tom_astrosource` aspect as well as setting up RTD and Travis.

The last 2 authors provided development input before I was involved, and before the code was in GitHub.
