### 2020-09-17 21aa223
Missing affiliation index for joe
