### 2020-09-25 4c66470
minor typo in g band filter lookup
