### 2020-09-27 7f2b82d
added nan value catching on opening photfiles
