### 2020-10-27 72de1de
catching an error where the variability array is empty
