+++ astrosource/plots.py
@@ -72,18 +72,19 @@ def open_photometry_files(outcatPath):
 def plot_variability(output, parentPath):
     # star Variability Plot
 
-    plt.cla()
-    outplotx = asarray(output)[:, 2]
-    outploty = asarray(output)[:, 3]
-    plt.xlabel('Mean Differential Magnitude of a Given Star')
-    plt.ylabel('Standard Deviation of Differential Magnitudes')
-    plt.plot(outplotx, outploty, 'bo')
-    # plt.plot(linex, liney)
-    plt.ylim(min(outploty)-0.04, max(outploty)+0.04, 'k-')
-    plt.xlim(min(outplotx)-0.1, max(outplotx)+0.1)
-    plt.grid(True)
-    plt.savefig(parentPath / 'starVariability.png')
-    plt.savefig(parentPath / 'starVariability.eps')
+    if output != []: # Do not attempt plot if output array is empty
+        plt.cla()
+        outplotx = asarray(output)[:, 2]
+        outploty = asarray(output)[:, 3]
+        plt.xlabel('Mean Differential Magnitude of a Given Star')
+        plt.ylabel('Standard Deviation of Differential Magnitudes')
+        plt.plot(outplotx, outploty, 'bo')
+        # plt.plot(linex, liney)
+        plt.ylim(min(outploty)-0.04, max(outploty)+0.04, 'k-')
+        plt.xlim(min(outplotx)-0.1, max(outplotx)+0.1)
+        plt.grid(True)
+        plt.savefig(parentPath / 'starVariability.png')
+        plt.savefig(parentPath / 'starVariability.eps')
     return
 
 
