### 2020-10-27 e42f5cb
added a bit of code to ignore zero-sized phot files
