### 2020-10-27 b9d3a3b
incorporated a routine that rejects probably erroneous catalogue magnitudes
