### 2020-10-27 25121fe
incorporated a routine that rejects probably erroneous catalogue magnitudes slight adjust
