### 2020-11-04 #2641 @joshspeagle
Review is done! Comments below:
- Documentation is functional enough to meet JOSS requirements I *think*, but definitely not super thorough. Docs seem more like they were written more as a reference and less as a guide for new users. There's also limited examples of the internal Python API that the CLI runs off of (a stated feature of the package), although there's enough stuff online and in the codebase for someone to figure it out by poking around. Adding those in would be helpful but not required.
- Automatic tests are helpful, although figuring out the outputs is nontrivial even with the explanations. Additional explicit guidance working through an example would help enormously here.
- No issues with the paper other than the figure appears to be mislabeled in the submitted PDF (it's "??" for me).
- While not a requirement for acceptance, I would encourage the author(s) to try and add more examples and documentation to make the code easier to work through. In particular, I think it's especially important to have worked-through cases illustrating for users how exactly statements like this actually work in the code: "Solutions to avoid these issues have been easily incorporated into the `astrosource` code, but would likely be cumbersome to explain manually in any given observer guide."

As most of the above issues are recommendations and not requirements, I'm happy to recommend the submission be accepted.
