### 2020-11-04 #2641 @zemogle
Thanks @joshspeagle Those are really constructive comments. The docs are an on-going effort but giving worked through examples are a good idea.

I've fixed up the figure references (and the missing captions).
