### 2020-11-17 fe600cc
caught empty VSX tables and a routine to catch RAand DEC for VSX around 0 RA
