### 2020-11-17 0fae587
a routine to catch RA and DEC for Calibration around 0 RA
