+++ astrosource/test/test_comparison.py
@@ -34,6 +34,11 @@ class TestSetup:
                 fid.write("{}\n".format(f))
         # Add targets to the TestSetup object
         self.targets = nparray([(2.92142, -1.74868, 0.00000000, 0.00000000)])
+        self.fileslist = [[f,0.,0.,0.] for f in TEST_PATHS['parent'].glob('*.npy')]
+
+        # self.photometry
+        # self.comparisons
+        # self.starvar
 
 @pytest.fixture()
 def setup():
@@ -41,9 +46,8 @@ def setup():
 
 def test_read_data_files(setup):
     files = os.listdir(TEST_PATHS['parent'])
-    fileslist = list(TEST_PATHS['parent'].glob('*.npy'))
     assert 'screenedComps.csv' in files
-    photFileArray, fileList = read_data_files(TEST_PATHS['parent'], fileslist)
+    photFileArray, fileList = read_data_files(TEST_PATHS['parent'], setup.fileslist)
     referenceFrame, rfid = find_reference_frame(photFileArray)
     assert list(referenceFrame[0]) == [154.7583434, -9.6660181000000005, 271.47230000000002, 23.331099999999999, 86656.100000000006, 319.22829999999999]
     assert rfid == 1
@@ -53,32 +57,33 @@ def test_read_data_files(setup):
 @patch('astrosource.comparison.Vizier.query_region',mock_vizier_query_region_vsx)
 def test_remove_targets_calibrated(setup):
     parentPath = TEST_PATHS['parent']
-    fileslist = TEST_PATHS['parent'].glob('*.npy')
-    photFileArray, fileList = read_data_files(parentPath, fileslist)
+    photFileArray, fileList = read_data_files(parentPath, setup.fileslist)
     assert photFileArray.shape[0] == 4
-    compFile_out = remove_stars_targets(photFileArray, acceptDistance=5.0, targetFile=setup.targets, removeTargets=1)
+    photometry = remove_stars_targets(photFileArray, acceptDistance=5.0, targetFile=setup.targets, removeTargets=1)
     # 3 stars are removed because they are variable
-    assert compFile_out.shape == (55,2)
+    assert photometry.shape == (55,2)
 
-@patch('astrosource.comparison.Vizier',mock_vizier_apass_b)
-def test_find_comparisons_calibrated_b(setup):
-    compFile = find_comparisons_calibrated(filterCode='B', paths=TEST_PATHS, targets=setup.targets, comparisons=,photometry=,starvar=)
-    assert compFile.shape == (10,5)
-
-@patch('astrosource.comparison.Vizier',mock_vizier_apass_v)
-def test_find_comparisons_calibrated_v(setup):
-    compFile = find_comparisons_calibrated(filterCode='V', paths=TEST_PATHS, targets=setup.targets, comparisons=,photometry=,starvar=)
-    assert compFile.shape == (10,5)
+# def test_calibrate_photometry(setup):
+#     calibrate_photometry(targets=setup.targets, filterCode='B', starvar, closerejectd, variabilityMultiplier
+#
+# @patch('astrosource.comparison.Vizier',mock_vizier_apass_b)
+# def test_find_comparisons_calibrated_b(setup):
+#     compFile = find_comparisons_calibrated(filterCode='B', paths=TEST_PATHS, targets=setup.targets, comparisons=,photometry=,starvar=)
+#     assert compFile.shape == (10,5)
+#
+# @patch('astrosource.comparison.Vizier',mock_vizier_apass_v)
+# def test_find_comparisons_calibrated_v(setup):
[…]
