+++ astrosource/test/test_identify.py
@@ -2,9 +2,10 @@ from astropy.io import fits
 import numpy
 import os
 from pathlib import Path
+import glob
 
 from astrosource.identify import (rename_data_file, export_photometry_files,
-    extract_photometry, gather_files, find_stars)
+    extract_photometry, gather_files, find_stars, read_data_files, find_reference_frame)
 
 
 TEST_PATH_PARENT = Path(os.path.dirname(__file__)) / 'test_files'
@@ -69,24 +70,46 @@ def test_extract_photometry(tmp_path):
     # Test if csv file is as we expect
     assert result_phot.all() == test_phot.all()
 
-def test_gather_files():
-
+def test_gather_files_fits():
     phot_files, filtercode = gather_files(TEST_PATHS, filetype="fits")
+    test_files = [['XOd2_ip_57757d0532642000_2017d01d04T01d16d43d571_1a089113_22d284_kb29.npy', 57757.0532642, 1.089113], ['XOd2_ip_57757d0522793000_2017d01d04T01d15d18d519_1a0899013_22d293_kb29.npy', 57757.0522793, 1.0899013]]
+    assert phot_files == test_files
+
+def test_gather_files_psx():
+    phot_files, filtercode = gather_files(TEST_PATHS, filetype="psx")
+    test_files = [['RXEri_B_2458371d6046093008_2018d09d10T02d21d20d615_1a1529288_17d283_kb96_michaelfitzgerald.npy', 2458371.604609301, 1.1529288], ['RXEri_B_2458371d6041340921_2018d09d10T02d20d40d435_1a1543828_17d283_kb96_michaelfitzgerald.npy', 2458371.604134092, 1.1543828]]
+    assert phot_files == test_files
+
+def test_find_reference_frames():
     test_files = [['XOd2_ip_57757d0532642000_2017d01d04T01d16d43d571_1a089113_22d284_kb29.npy',0.0,0.0], ['XOd2_ip_57757d0522793000_2017d01d04T01d15d18d519_1a0899013_22d293_kb29.npy',0.0,0.0]]
-    assert phot_files.sort() == test_files.sort()
+    photFileArray, fileList = read_data_files(TEST_PATHS['parent'], test_files)
+    print(photFileArray.shape, len(fileList))
+    referenceFrame, rfid = find_reference_frame(photFileArray)
+    assert rfid == 0
 
 def test_find_stars():
     targets = numpy.array([[117.0269708, 50.2258111, 0,0]])
     phot_files, filtercode = gather_files(TEST_PATHS, filetype="fits")
-    usedImages = find_stars(targets, TEST_PATHS, phot_files)
-    images_list = [str(u) for u in usedImages]
+    usedImages, comparisons, targetphot = find_stars(targets, TEST_PATHS, phot_files)
+    print(usedImages)
+    print(comparisons)
+    print(targetphot)
+    images_list = [str(u[0]) for u in usedImages]
     # Check the right files are saved
     test_list = (TEST_PATHS['parent'] / 'usedImages_test.txt').read_text().strip().split('\n')
-    assert images_list.sort() == test_list.sort()
+    assert images_list == test_list
+
+def test_read_data_files():
+    test_files = [['XOd2_ip_57757d0532642000_2017d01d04T01d16d43d571_1a089113_22d284_kb29.npy',0.0,0.0], ['XOd2_ip_57757d0522793000_2017d01d04T01d15d18d519_1a0899013_22d293_kb29.npy',0.0,0.0]]
+    photarray, filelist = read_data_files(TEST_PATHS['parent'], test_files)
+    assert photarray.shape == (2,)
+    assert len(filelist) == 2
+
+def tear_down():
[…]

[…]
