### 2021-01-26 #2641 @arfon
:wave: @gusbeane @zackdraper - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under review here is [_astrosource: automating optical astronomy measurement, calibration and analysis for variable stellar sources from provided photometry_](https://github.com/zemogle/astrosource/) which already has one complete review and we're looking for a second reviewer to replace @bsipocz who unfortunately isn't able to complete their review at this time.
