### 2021-01-28 #2641 @zackdraper
Sure, I'll give it a go over the weekend.  Haven't done a JOSS review before but it looks straightforward.
