### 2021-01-28 #2641 @arfon
> Sure, I'll give it a go over the weekend. Haven't done a JOSS review before but it looks straightforward.

⚡ thank you so much @zackdraper! Please review [the instructions above](https://github.com/openjournals/joss-reviews/issues/2641#issuecomment-687649852), and work through the checklist at the top of the issue (you'll need to accept the repository invite above from Whedon before you can update the checklist).

Any questions, don't hesitate to ask!
