### 2021-02-03 #2641 @zackdraper
@zemogle Do you have a link to an example data set to use?  Trying to confirm functionality on real data. Tried getting public data off of LCO and got a few errors.

The error for a missing keyword was caught, but I would suggest having documentation saying what fits header keywords are required for the input data.

I got another error resulting from a malformatted .fits file on "hdulist[2]" giving a Python KeyError.
