### 2021-02-03 #2641 @zemogle
Hey @zackdraper 👋  Thanks for picking this up. Just picking a random data set, you could try this data set of [Tres 2b](https://archive.lco.global/?q=a&RLEVEL=91&PROPID=&INSTRUME=&OBJECT=&SITEID=&TELID=&FILTER=&OBSTYPE=&EXPTIME=&BLKUID=&REQNUM=1849306&basename=&start=2014-01-01&end=2021-07-31%2023%3A59&id=&public=true).

The coords of Tres 2b are `--ra 286.8083 --dec 49.3164` if you need them. I've just just tried `astrosource` with this dataset using the `--full` option and left it. All worked as expected.

I think the error you mentioned might have been from a fits file which doesn't have the photometry data in the fits extension.

Good catch on the documentation. As with all documentation it is a work in progress. I've never written documentation before, so any and all suggestions are helpful.
