### 2021-02-11 #2641 @zackdraper
Hey @zemogle I ran it on the Tres2b data and got an error, looks like there needs to be more exception handling when intermediate steps don't return what is normally expected.

```
(py3) zackdraper@zackdraper-MacBookAir:~/JOSS$ astrosource --ra 286.8083 --dec 49.3164 --indir lcogtdata-20210211-10/ --full
💾 Inspecting input files
🌟 Identify comparison stars for photometry calculations
..........
⭐️ Find stable comparison stars for differential photometry
..................................................................................................................................................................................................💫.................................................................................................................................................................................................💫...............................................................................................................................................................................................
⭐️ Find comparison stars in catalogues for calibrated photometry
..........
Traceback (most recent call last):
  File "/home/zackdraper/anaconda2/envs/py3/bin/astrosource", line 11, in <module>
    sys.exit(main())
  File "/home/zackdraper/anaconda2/envs/py3/lib/python3.6/site-packages/click/core.py", line 722, in __call__
    return self.main(*args, **kwargs)
  File "/home/zackdraper/anaconda2/envs/py3/lib/python3.6/site-packages/click/core.py", line 697, in main
    rv = self.invoke(ctx)
  File "/home/zackdraper/anaconda2/envs/py3/lib/python3.6/site-packages/click/core.py", line 895, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "/home/zackdraper/anaconda2/envs/py3/lib/python3.6/site-packages/click/core.py", line 535, in invoke
    return callback(*args, **kwargs)
  File "/home/zackdraper/anaconda2/envs/py3/lib/python3.6/site-packages/astrosource/main.py", line 90, in main
    ts.find_variables()
  File "/home/zackdraper/anaconda2/envs/py3/lib/python3.6/site-packages/astrosource/astrosource.py", line 69, in find_variables
    find_variable_stars(targets=self.targets, parentPath=self.paths['parent'])
  File "/home/zackdraper/anaconda2/envs/py3/lib/python3.6/site-packages/astrosource/analyse.py", line 175, in find_variable_stars
    plot_variability(outputVariableHolder, parentPath)
  File "/home/zackdraper/anaconda2/envs/py3/lib/python3.6/site-packages/astrosource/plots.py", line 74, in plot_variability
    outplotx = asarray(output)[:, 2]
IndexError: too many indices for array: array is 1-dimensional, but 2 were indexed

```
