### 2021-02-11 #2641 @zemogle
Hmmm. This is puzzling. I just made a new environment and installed astrosource from `pypi`. Everything ran smoothly. I also did a checkout of `master` to see if there was an issue there, but it also ran smoothly. None of our other users have reported this problem either so I'm finding it a bit tricky to debug.

Would you be able to make an [issue](https://github.com/zemogle/astrosource/issues)  so we keep this separate from the paper review ?
