### 2021-02-12 #2641 @zackdraper
I have to verify the code actually works.  Do you have another data set of an actual variable star?  I think at the very least you would need to document an example dataset people can use. Obviously cant include it because the data would be too large, but some public data source is useful for this code to be tested with.  Also, the fits files require certain keywords to work, I think you need to document those requirements as well.  Imagine someone tries this on another observatory dataset and it happens to use non-standard fits headers. This would make it more useful for the amateur's you are trying to help with this software.

- Functionality: Have the functional claims of the software been confirmed?
- Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I followed a fresh install as documented, but I'll try it on another computer just to be sure.

I appreciate that this review process has gone on for a very long time for you and hope to get it finished soon.
