### 2021-02-12 #2641 @zemogle
I think the docs should be clearer on this but `astrosource` doesn't do the photometric extraction, it just does the photometric analysis. It really only operates on FITS files which have a photometry table extension. So the inclusion of FITS format is just for data tables not images (see [Basic Usage](https://astrosource.readthedocs.io/en/latest/basic_usage.html#time-series-data) part of the docs). The exact columns needed are also detailed in that section.

Here is a small public dataset of variable star [RX Eri](https://cloud.lco.global/s/xS2ZLkLsDyJyagx). You'll need to use the following options with it because these are `.psx` format not `FITS Table`.
```
--ra 72.43451
--dec -15.74109
--format psx
```

If that doesn't work, I suspect there is something about the versions of `NumPy` or `Matplotlib` which is causing the shenanigens. It would be really helpful if you could post an issue, so we can track the problem separately to the paper review. As I understand it the paper review should be higher level than code bugs.

Thanks for bearing with us!
