### 2021-02-17 #2641 @zackdraper
Ok, I'm satisfied that the code is functional now with that example.  But, I would strongly encourage you to include an actual example like the RX Eri data set somewhere in the documentation.  You have a few examples shown in the paper. Ideally, for open source software, I would think those examples should be readily available to all users to immitate and prepare thier own data. `astrosource --help` also needs better documentation on what the parameters are (for ease of usabaility, the parameters are doccumented suffeciently in the readme).  The exception handling issues can be chalked up to bugs to be fixed, but usable examples are really how others are going to take this open source code as a starting point to do thier own work, modify the code, etc.
