### 2021-02-17 #2641 @zemogle
Thanks @zackdraper  and @joshspeagle for very constructive reviews. We'll find a way to include a useful dataset and add more complete examples to the docs 👍
