+++ astrosource/main.py
@@ -13,46 +13,46 @@ from astrosource.utils import get_targets, folder_setup, AstrosourceException, c
 logger = logging.getLogger('astrosource')
 
 @click.command()
-@click.option('--full', is_flag=True)
-@click.option('--stars', is_flag=True)
-@click.option('--comparison', is_flag=True)
-@click.option('--calc', is_flag=True)
-@click.option('--calib', is_flag=True)
-@click.option('--phot', is_flag=True)
-@click.option('--plot', is_flag=True)
-@click.option('--detrend', is_flag=True)
-@click.option('--eebls', is_flag=True)
-@click.option('--period', is_flag=True, type=float)
-@click.option('--indir', default=None, type=str, required=True)
-@click.option('--ra', type=float)
-@click.option('--dec', type=float)
-@click.option('--target-file', default=None, type=str)
-@click.option('--format', default='fz', type=str)
-@click.option('--imgreject', '-ir', type=float, default=0.05)
-@click.option('--bjd', is_flag=True)
-@click.option('--clean', is_flag=True)
-@click.option('--verbose', '-v', is_flag=True)
-@click.option('--periodlower', '-pl', type=float, default=-99.9)
-@click.option('--periodupper', '-pu', type=float, default=-99.9)
-@click.option('--periodtests', '-pt', type=int, default=-99)
-@click.option('--rejectbrighter', '-rb', type=float, default=99)
-@click.option('--rejectdimmer', '-rd', type=float, default=99)
-@click.option('--thresholdcounts', '-tc', type=int, default=1000000)
-@click.option('--hicounts', '-hc', type=int, default=1500000)
-@click.option('--lowcounts',  '-lc', type=int, default=1000)
-@click.option('--starreject', '-sr', type=float, default=0.3)
-@click.option('--closerejectd', '-sr', type=float, default=5.0)
-@click.option('--nopanstarrs', '-np', is_flag=True)
-@click.option('--mincompstars', '-mc', type=float, default=0.1)
-@click.option('--nosdss', '-ns', is_flag=True)
-@click.option('--skipvarsearch', '-sv', is_flag=True)
+@click.option('--full', is_flag=True, help='Perform full analysis, i.e. steps 1-5')
+@click.option('--stars', is_flag=True, help='Step 1: Identify and match stars from each data file')
+@click.option('--comparison', is_flag=True, help='Step 2: Identify non-varying stars to use for comparisons')
+@click.option('--calc', is_flag=True, help='Step 3: Calculate the brightness change of the target')
+@click.option('--phot', is_flag=True, help='Step 4: Photometry calculations for either differential or calibrated')
+@click.option('--plot', is_flag=True, help='Step 5: Produce lightcurve plots')
+@click.option('--calib', is_flag=True, help='Perform calibrations for absolute photometry')
+@click.option('--detrend', is_flag=True, help='Detrend exoplanet data')
+@click.option('--eebls', is_flag=True, help='Box fitting to search for periodic transits')
+@click.option('--indir', default=None, type=str, required=True, help='Path of directory containing LCO data files. If none is given, astrosource assumes the current directory')
+@click.option('--ra', type=float, help='Right Ascension of the target (in decimal)')
+@click.option('--dec', type=float, help='Declination of the target (in decimal)')
+@click.option('--target-file', default=None, type=str, help='Textfile in csv with RA and Dec in decimal')
+@click.option('--format', default='fz', type=str, help='Input file format. If not `fz`, `fits`, or `fit` assumes the input files are photometry files with correct headers. If image files given, code looks for photometry in FITS Table extension.')
+@click.option('--imgreject', '-ir', type=float, default=0.05, help=' Image fraction rejection allowance based on image size starting value.')
[…]
