+++ docs/source/basic_usage.rst
@@ -66,15 +66,15 @@ The above options will perform an basic analysis of the provided time-series fil
 **detrend** `boolean flag`
   Detrend exoplanet data
 **period** `boolean flag`
-  Search for periodicity in the data. 
+  Search for periodicity in the data.
 **eebls** `boolean flag`
   EEBLS - box fitting to search for periodic transits
 **calib** `boolean flag`
   Perform calibrated photometry to get data in absolute magnitudes
 **clean** `boolean flag`
   Remove all files except the original data files
-  
-**periodlower** `float` 
+
+**periodlower** `float`
   Shortest period to trial in days. Default is 0.05
 
 **periodupper** `float`
@@ -84,7 +84,7 @@ The above options will perform an basic analysis of the provided time-series fil
   Number of different trial periods to run, Default 10000
 
 **skipvarsearch** `boolean flag`
-  If this is set, this skips the variability calculations for each identified star. Pragmatically this skips creating the starVariability outputs. In a crowded field this can take an excessive amount of time. 
+  If this is set, this skips the variability calculations for each identified star. Pragmatically this skips creating the starVariability outputs. In a crowded field this can take an excessive amount of time.
 
 **lowcounts** `int`
   Countrate above which to accept a comparison star as a candidate. Defaults to 1000. (Note: This may seem like a low number but realistically stars at this low rate will be highly variable and rejected anyway)
@@ -93,7 +93,7 @@ The above options will perform an basic analysis of the provided time-series fil
   Countrate above which to reject a comparison star as a candidate. Defaults to 1000000. (Note: This will vary from camera to camera, but it should be representative of a typical value that would have a peak pixel value significantly below the full range of the ccd, preferably lower rather than higher. )
 
 **thresholdcounts** `int`
-  the number of counts at which to stop adding identified comparison stars to the ensemble. Default 1000000. 
+  the number of counts at which to stop adding identified comparison stars to the ensemble. Default 1000000.
 
 **closerejectd** `float`
   astrosource will reject potential comparison calibration stars if there are nearby stars to it. Closerejectd is the limiting distance in arcseconds. For crowded fields, this value may need to be lowered. Default 5.0. While the primary function here is to identify stars that are not in crowded situations, it also has the side-effect of removing false detections in sky surveys due to image artifacts or diffraction spikes which tend to cluster together.
@@ -101,7 +101,7 @@ The above options will perform an basic analysis of the provided time-series fil
 **nopanstarrs** `boolean flag`
   Do not use the PanSTARRS catalogue for calibration. Some regions of the sky in PanSTARRS have poorer quality photometry than SDSS.
 
-**sdss** `boolean flag`
+**nosdss** `boolean flag`
   Do not use the SDSS catalogue for calibration. Some regions of the sky in SDSS have poorer quality photometry than PanSTARRS.
 
 **bjd** `boolean flag`
@@ -146,9 +146,9 @@ A variety of output files are generated. Some are obvious, some are not so obvio
 
 **calibStands.csv**: There are stars from stdComps.csv (as opposed to compsUsed.csv) that have identified calibrated magnitudes in the catalogue. The magnitudes in this catalogue are from the reference catalogue (e.g. APASS, Skymapper, SDSS or PanSTARRS)
 
-**calibCompsUsed.csv**: These are the calibrated magnitudes of the compsUsed.csv comparison stars calculated by comparison to calibStands.csv. 
+**calibCompsUsed.csv**: These are the calibrated magnitudes of the compsUsed.csv comparison stars calculated by comparison to calibStands.csv.
 
[…]

[…]
