+++ .github/workflows/publish-to-test-pypi.yml
@@ -7,7 +7,7 @@ jobs:
     name: Build and publish astrosource package 📦 to PyPI and TestPyPI
     runs-on: ubuntu-18.04
     steps:
-    - uses: actions/checkout@master
+    - uses: actions/checkout@main
     - name: Set up Python 3.7
       uses: actions/setup-python@v1
       with:
