### 2021-02-25 #2641 @zemogle
I've updated the code with the suggestions, published on PyPi and made a release which is on Figshare. Here is the [DOI](https://doi.org/10.6084/m9.figshare.14113667)

I think that's it! Let me know if need anything else for publication. This has been a really great experience! 👍
