### 2021-02-26 #2641 @kthyng
Huh, I don't see figshare come through too often so I might be looking in the wrong place, but I am having trouble verifying that the title and author list match the JOSS submission. @zemogle can you point me to where this is, or update it on your end?
