### 2021-02-26 #2641 @zemogle
@kthyng If you click the DOI link you should be able to see the title and authors list (Figshare shows a preview of the archive first, bizarrely).

The version matching the paper is 1.5.2
